SheQuLeSong/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
tortoise-orm==0.17.7
passlib==1.7.4
PyJWT==2.3.0
python-multipart==0.0.5
bcrypt==3.2.0
aiosqlite==0.17.0
python-dateutil==2.8.2
aiofiles==0.8.0 